ci(catalog): Improvement to unittest cases in catalog/registry.go (#264)
Hi team,

I'm new to iceberg and iceberg-go and working on the adoption at my
current company. I was digging at the code and found some area to
contribute to. Not saying achiving 100% code coverage is always good but
it serves as an entry point for my understanding of the intenals :bow:

![image](https://github.com/user-attachments/assets/d519275f-0f4e-46fd-b1cd-66219df3843c)

---------

Signed-off-by: dttung2905 <ttdao.2015@accountancy.smu.edu.sg>