refactor(catalog): restructure catalog package (#266)
Refactoring and restructuring the `catalog` package based on discussions
with others in the Go community. This achieves the following benefits:

* Improved, and shorter, naming of types and methods. Things that are
specific to the REST catalog or Glue catalog can be put directly into
those packages allowing for a public API such as `rest.NewCatalog`
instead of `catalog.NewRestCatalog`. Options can be made specific to
catalog implementations without weird generic handling, etc.
* Using the Registry allows consumers to only import the catalog
implementations they care about, allowing any others to be pruned and
removed by the compiler as unused and unreachable if not imported. This
will enable smaller binaries if not all catalog implementations are
needed.