feat(catalog): Initial implementation of sql catalog (#246)
Building on #266 (which would need to be merged before this), this
creates an initial implementation of a SQL catalog.

We're utilizing the https://bun.uptrace.dev/ ORM to handle dispatching
queries among SQLite, Postgres, Oracle, MySQL, and MSSQL dialects, but
accepts a standard `*sql.DB` object and custom drivers rather than
bundling them directly into this package. That provides extra
flexibility as long as the desired database supports one of the SQL
dialects that are implemented.