fix(table/scanner): Fix nested field scan (#311)
Fixes #309 

There was a combination of factors that caused the initial problem:

1. The arrow-go/v18/parquet/pqarrow library wasn't properly propagating
`PARQUET:field_id` metadata for children of List or Map typed fields
2. We only iterated the *fields* and skipped list/maptypes when
selecting column indexes, this caused us to miss the children. Instead
we need to iterate all of the *field IDs*, this change updates that.
3. When pruning parquet fields we were not propagating the correct
ColIndex for map typed columns, we want the leaves so we need the
ColIndex of the children
4. creating the output arrays during `ToRequestedSchema` led to a memory
leak for list/map columns that needed to be fixed.

A unit test has been added to ensure we are properly able to read the
`test_all_types` table and get the rows without error.