ci(golangci-lint) :Add golangci linter (#315)
From the feedback
https://github.com/apache/iceberg-go/pull/306#discussion_r1960205322,
I'm adding golangci-lint in GIthub Action CI. There are many different
linters but I'm just adding the most basic one. We can definitely add
more if needed

I have not run golangci-lint with `--fix` argument. I plan to get it
flagged out in the PR as a form of testing. Let me know if we want to
commit the changes after fixing

---------

Signed-off-by: dttung2905 <ttdao.2015@accountancy.smu.edu.sg>