feat(table): Add computation of iceberg stats from parquet files (#329)
This one's a big one!

Adding initial implementation of deriving and computing the Iceberg
metadata statistics from Parquet files in order to facilitate adding
files directly. This includes utilizing the Parquet file metadata
retrieve the typed statistics and collecting them based on optional
metrics modes and so on. This also includes various unit tests to
confirm it is working.