feat: (catalog/glue) Add support for CreateTable (#326)
Hi team,
This PR aims to support CreateTable for glue catalog. Below are the list
(I think) to be done:
- [x] Tested out on a real Glue Catalog .Table was created successfully
- [x] Add integration test to this new method. I have been struggling to
create a unit test for it because after the table is created, we need to
load the table state and return it. At this loading step, a call to s3
to check for the metadata file existence was made by `LoadTable()`.
Hence, I'm taking reference from `TestGlueLoadTableIntegration` and
create a few integration test cases. Ran the test successfully with a
real table in AWS Glue by setting `TEST_DATABASE_NAME` ,
`TEST_TABLE_NAME` and `TEST_TABLE_LOCATION` env variables to an existing
one in Glue Catalog

---------

Signed-off-by: dttung2905 <ttdao.2015@accountancy.smu.edu.sg>