feat(table, catalog/sql): Implement CommitTable and connect it with transactions (#339)
With this, we can officially add files and commit the transaction to
catalogs. I've tested this and the committed table is able to be read
and scanned by pyiceberg!

Later on I'll look into creating a more solid integration test directly
with pyiceberg, but for now this completes the fully implementation of
AddFiles! (albeit, we still don't handle failures gracefully yet, but
that's a problem for future-me)

closes #287