feat(table): write an Arrow Table or stream of Arrow records (#369)
This creates the initial implementation to write a stream of data or a
materialized Table of arrow data to unpartitioned tables. Implementing
writing with partitions will be done in a follow-up PR.

For now this will only write Parquet files, but everything is abstracted
well enough to easily allow for other file types to be implemented.