fix(catalog/rest): Fix concurrency bug in REST catalog request signing (#384)
A hasher is not thread-safe, yet the same hasher was being used for all
requests. If applications made concurrent calls to the same REST catalog
implementation, they could end up writing to the same hasher, corrupting
the signatures for both concurrent requests.

This makes things thread-safe by creating a hasher for each signing
operation. The could be safely re-used using a `sync.Pool`. But the
hasher is only 116 bytes, and initialization just has to write 8 bytes
(other than the zero'ing done by the allocator), so it doesn't seem
worth trying to re-use them.