fix(table): ensure manifest length matches file size (#439)
fixes #438

We need to ensure that we flush the manifest that we're writing before
we construct the `ManifestFile` object we want to add to the snapshot so
that our length is correct when we write the manifest list itself. This
change adds unit tests to ensure that the lengths match up correctly
with what we write, and also ensures the tests work on windows (ran into
some problems when putting this PR together that I had to fix).