fix(table): fix race condition in scanner (#445)
fixes #443

Eliminates potential race condition in scanning caused by the metrics
evaluator used for manifest pruning.

Since the stats are only used for a given run of Eval, we don't need to
modify the bound version of the `inclusiveMetricsEval` object and can
instead copy the type and expression (cheap) to populate for the call.
This avoids the race condition without giving up performance (no locking
needed).