fix(schema): Fix bug in Schema.HighestFieldId (#494)
## Description
This pull request fixes a bug in computing the HighestFieldID when a
schema contains map types. The `findLastFieldID` visitor previously
would return the max of the key result and value result, while it should
also be taking the max of it's `KeyID` and `ValueID`.

## Testing
This PR introduced a new `MapType` field on the test schema, and a new
test case, `TestHighestFieldID` to validate the computation. It's worth
noting that this edge case only gets hit when a map of primitive types
is the last field in the schema.

The test case was run without and with the bug fix to validate the
changes.