feat(table): Implement snapshot expiration (#401)
Hi @zeroshade,

As discussed on a previous PR, I'm trying to implement snapshots
expiration now.
This is not complete, but I wanted to discuss it with you before I
further work on it.


If the design looks ok to you, here are a few technical questions
regarding snapshot expiration:
- currently, the snapshots field of the produced metadata file looks ok
(expired snapshots are not there), but the snapshot-log field still
contains entries for every operation since table creation. Is is
expected ? I don't find an appropriate answer in the spec.
- should we handle expired data and metadata files deletion here ?
- the produced metadata file does not contains any new snapshot (not
snapshot is created for the expire snapshots operation). I cannot find
out what the spec says about it. Most probably a new snapshot must be
created.

## References

- https://iceberg.apache.org/spec/#snapshot-retention-policy