fix(manifest): Interpret manifest files written without `content` metadata as `data` files (#545)
fixes: #544 

V1 writers may not specify `content` at all when writing manifest files,
but since this predates `delete` files these should be interpretted as
`data` files
([https://iceberg.apache.org/spec/#manifests](https://iceberg.apache.org/spec/#manifests)).

_Side Note: It could be helpful to have a `iceberg-testing` repo similar
to `parquet-testing` or `arrow-testing` that contains a bunch of files
serialized to different standards of the spec. I don't know if this
exists already but it was a bit tricky producing a manifest file to use
for a unit test since the V1 writer included writes `content=data` for
compatibility. I started creating new writer that omits that one field
but it seemed like more to maintain than it was worth for that one
test._

---------

Co-authored-by: Joel Lubinitsky <jlubinitsky@netflix.com>