refactor(catalog): remove `props` parameter from LoadTable and change `table.Table` to `table.Identifier` in CommitTable (#542)
### PR Description

In our team's usage of the `Catalog` interface, we found the `LoadTable`
and `CommitTable` methods to be somewhat confusing.

1. **LoadTable**: The `props` parameter was included in the interface,
but it was not actually used. All IO-related properties are already
encapsulated within the catalog's own properties. In practice, this
parameter was often set to `nil`, which made its presence unnecessary.
Notably, in the Python, Rust, and Java implementations, the `LoadTable`
method does not include the `props` parameter at all.

2. **CommitTable**: Across all catalog implementations, including
`rest`, `glue`, and `sql`, the `tbl.Identifier()` was consistently used.
When calculating the `StageTable`, the current table was reloaded using
`LoadTable`, yet the parameter required was a `table.Table`. This
inconsistency prompted the need for change. In the Rust and Java
implementations, `CommitTable` already uses `table.Identifier`. In the
Python implementation, while it uses `table.Table`, it similarly does
not utilize any methods other than `tbl.Identifier()`.

To address these issues, I have made the following changes in this PR:
- Removed the `props` parameter from the `LoadTable` method.
- Changed the `table.Table` parameter to `table.Identifier` in the
`CommitTable` method.

If there are any misunderstandings or if I've missed something, please
let me know so we can clarify or close this PR.