fix(schema): Fix bug in Schema.HighestFieldId for List types (#611)
## Description
This pull request fixes a bug in computing the `HighestFieldID` when a
schema contains list types. The `findLastFieldID` visitor previously
would only return the computed highest field of a list's element.
Instead, it should return the max of the `ElementID` and the computed
highest field ID of the element field.

Since primitive types trivially have zero-valued field IDs, this would
always return zero for those types. Even if the type were non-primitive,
it would break if IDs were not assigned using preorder traversal (though
of course preorder is recommended).

## Testing
Added a new test case in `schema_test.go` to validate the minimum
reproducible example for this bug.