feat(view): Implement view metadata builder and UpdateView method in REST catalog (#629)
## Description
This PR implements full support for updating Iceberg views for the REST
catalog

## Changes
- Implements an `UpdateView` method on the REST catalog
- Implements `view.MetadataBuilder` to generate updates for use in
`UpdateView` on the REST catalog
- I did my best to blend the style of the existing codebase and use
idiomatic Go, while still being faithful to the Java reference code.
Some of the refactors of existing code bring the state of the view
package more inline with Java
- I transcribed every applicable unit test from the Java
`ViewMetadataBuilder` tests to validate behavior
- Refactors the existing state of the `view` package substantially
- I apologize for the diff here, I was working off of a fork for a few
weeks and had not previously seen the recent additions to view metadata
support. Did my best to reconcile them
- Refactors the REST catalog `CreateView` method to take variadic
`CreateViewOpt`s and now return a `view.View`, similar to the behavior
of the `CreateTable` method
- Added unit and integration tests to the `rest` package to cover the
new changes to the REST view methods
- Small refactor to extract `sliceEqualHelper` from `table/metadata.go`
to `internal/utils.go` for use in the `view` package
- Refactors the CreateMetadata function used by the SQL catalog to
return a `*View`, and extract the location from it