feat(metadata): Add support for source-ids (#651)
Part of #589 

We don't currently support the `source-ids` field on PartitionFields.
This will be more important when we have support for multi-arg
transforms.

Since we don't have support for them, I'm just using the UnmarshalJSON
field to set SourceID. We'll need a (much larger) refactor eventually to
support an array of SourceIDs.